fix: 앨범 페이지 Suspense 적용으로 빌드 에러 수정
